requirements.txt
psycopg2-binary
kafka-python